Switch to bash and fix shellcheck warnings. (#3)
Resource scripts use features that are not standard in posix `sh`, such
as `[[]]` tests. Also, the `sh` that ships with alpine treats
`build_urls` as empty when not quoted in `[[ ! -z $build_urls ]]`.
Rather than handling edge cases in alpine `sh`, this patch switches to
`bash` and incidentally fixes some shellcheck warnings.